more minor css changes, some component renaming
