+++ src/components/HeaderBar/CreateModal.js
@@ -14,25 +14,13 @@ function CreateModal({ closeModal }) {
     <div className='modal-container'>
       <p>What would you like to create?</p>
       <div className='buttons'>
-        <tbody>
-          <tr>
-            <button onClick={handleClickGrocery} className='list-btn'>
-              Grocery List
-            </button>
-          </tr>
-          <tr>
-            <button className='recipe-btn'>
-              Recipe
-            </button>
-          </tr>
-          {/* <tr>
-            <button onClick={() => closeModal(false)} className='close-btn'>
-              Close
-            </button>
-          </tr> */}
-        </tbody>
-        
-        
+          <button onClick={handleClickGrocery} className='list-btn'>
+            Grocery List
+          </button>
+          
+          <button className='recipe-btn'>
+            Recipe
+          </button>        
       </div>
     </div>
