+++ src/components/HeaderBar/HomeModal.css
@@ -1,9 +1,9 @@
 .modal{
     position: absolute;
-    top: 110px;
-    right: 400px;
+    top: 105px;
+    right: 380px;
     width: 190px;
-    height: 310px;
+    height: 290px;
     border-radius: 12px;
     background-color: white;
     box-shadow: rgba(0, 0, 0, 0.35) 0px 5px 15px;
@@ -17,7 +17,7 @@
 
 .buttons {
     display: flex;
-    flex-direction: row;
+    flex-direction: column;
     justify-content: space-evenly;  
 }
 
@@ -49,7 +49,7 @@
 
 .kale-sm img {
     margin-right: 10px;
-    margin-top: 20px;
+    margin-top: -10px;
 }
 
 .close-btn2 {
